Rename CardReaderConnectFragment to CardReaderDetailFragment
