+++ WooCommerce/src/main/kotlin/com/woocommerce/android/ui/prefs/cardreader/detail/CardReaderDetailFragment.kt
@@ -1,4 +1,4 @@
-package com.woocommerce.android.ui.prefs.cardreader.connect
+package com.woocommerce.android.ui.prefs.cardreader.detail
 
 import android.os.Bundle
 import android.view.View
@@ -6,27 +6,27 @@ import androidx.fragment.app.viewModels
 import androidx.lifecycle.Observer
 import androidx.navigation.fragment.findNavController
 import com.woocommerce.android.R
-import com.woocommerce.android.databinding.FragmentSettingsCardReaderConnectBinding
+import com.woocommerce.android.databinding.CardReaderDetailBinding
 import com.woocommerce.android.extensions.navigateSafely
 import com.woocommerce.android.ui.base.BaseFragment
-import com.woocommerce.android.ui.prefs.cardreader.connect.CardReaderConnectViewModel.NavigationTarget.CardReaderScanScreen
+import com.woocommerce.android.ui.prefs.cardreader.detail.CardReaderDetailViewModel.NavigationTarget.CardReaderScanScreen
 import dagger.hilt.android.AndroidEntryPoint
 
 @AndroidEntryPoint
-class CardReaderConnectFragment : BaseFragment(R.layout.fragment_settings_card_reader_connect) {
-    val viewModel: CardReaderConnectViewModel by viewModels()
+class CardReaderDetailFragment : BaseFragment(R.layout.card_reader_detail) {
+    val viewModel: CardReaderDetailViewModel by viewModels()
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
 
-        val binding = FragmentSettingsCardReaderConnectBinding.bind(view)
+        val binding = CardReaderDetailBinding.bind(view)
 
         initViews(binding)
         initObservers()
     }
 
-    private fun initViews(binding: FragmentSettingsCardReaderConnectBinding) {
-        binding.testingText.setText("Hardcoded: Card Reader Connect Fragment")
+    private fun initViews(binding: CardReaderDetailBinding) {
+        binding.testingText.setText("Hardcoded: Card Reader Detail Fragment")
         binding.initiateScanBtn.setOnClickListener {
             viewModel.onInitiateScanBtnClicked()
         }
@@ -36,7 +36,7 @@ class CardReaderConnectFragment : BaseFragment(R.layout.fragment_settings_card_r
         viewModel.event.observe(viewLifecycleOwner, Observer {
             when (it) {
                 is CardReaderScanScreen ->
-                    findNavController().navigateSafely(R.id.action_cardReaderConnectFragment_to_cardReaderScanFragment)
+                    findNavController().navigateSafely(R.id.action_cardReaderDetailFragment_to_cardReaderScanFragment)
                 else ->
                     it.isHandled = false
             }
