+++ WooCommerce/src/main/res/layout/card_reader_detail.xml
