+++ WooCommerce/src/main/res/layout/fragment_settings_card_reader.xml
@@ -30,17 +30,17 @@
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
         android:text="Use simulated card reader"
-        app:layout_constraintBottom_toTopOf="@+id/redirect_to_connect_fragment"
+        app:layout_constraintBottom_toTopOf="@+id/redirect_to_detail_fragment"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toBottomOf="@id/connection_status"
         tools:ignore="HardcodedText" />
 
     <com.google.android.material.button.MaterialButton
-        android:id="@+id/redirect_to_connect_fragment"
+        android:id="@+id/redirect_to_detail_fragment"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:text="Redirect to ConnectFragment"
+        android:text="Redirect to CardReaderDetailFragment"
         app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
