+++ WooCommerce/src/main/res/navigation/nav_graph_settings.xml
@@ -57,11 +57,11 @@
         android:name="com.woocommerce.android.ui.prefs.cardreader.scan.CardReaderScanFragment"
         android:label="CardReaderScanFragment" />
     <fragment
-        android:id="@+id/cardReaderConnectFragment"
-        android:name="com.woocommerce.android.ui.prefs.cardreader.connect.CardReaderConnectFragment"
-        android:label="CardReaderConnectFragment" >
+        android:id="@+id/cardReaderDetailFragment"
+        android:name="com.woocommerce.android.ui.prefs.cardreader.detail.CardReaderDetailFragment"
+        android:label="CardReaderDetailFragment" >
         <action
-            android:id="@+id/action_cardReaderConnectFragment_to_cardReaderScanFragment"
+            android:id="@+id/action_cardReaderDetailFragment_to_cardReaderScanFragment"
             app:destination="@id/cardReaderScanFragment"
             app:enterAnim="@anim/activity_slide_in_from_right"
             app:exitAnim="@anim/activity_slide_out_to_left"
@@ -75,8 +75,8 @@
         android:name="com.woocommerce.android.ui.prefs.CardReaderSettingsFragment"
         android:label="CardReaderSettingsFragment" >
         <action
-            android:id="@+id/action_cardReaderSettingsFragment_to_cardReaderConnectFragment"
-            app:destination="@id/cardReaderConnectFragment"
+            android:id="@+id/action_cardReaderSettingsFragment_to_cardReaderDetailFragment"
+            app:destination="@id/cardReaderDetailFragment"
             app:enterAnim="@anim/activity_slide_in_from_right"
             app:exitAnim="@anim/activity_slide_out_to_left"
             app:popEnterAnim="@anim/activity_slide_in_from_left"
