+++ WooCommerce/src/test/kotlin/com/woocommerce/android/ui/prefs/cardreader/detail/CardReaderDetailViewModelTest.kt
@@ -1,7 +1,7 @@
-package com.woocommerce.android.ui.prefs.cardreader.connect
+package com.woocommerce.android.ui.prefs.cardreader.detail
 
 import androidx.lifecycle.SavedStateHandle
-import com.woocommerce.android.ui.prefs.cardreader.connect.CardReaderConnectViewModel.NavigationTarget
+import com.woocommerce.android.ui.prefs.cardreader.detail.CardReaderDetailViewModel.NavigationTarget
 import com.woocommerce.android.util.CoroutineTestRule
 import com.woocommerce.android.viewmodel.BaseUnitTest
 import kotlinx.coroutines.ExperimentalCoroutinesApi
@@ -11,15 +11,15 @@ import org.junit.Rule
 import org.junit.Test
 
 @ExperimentalCoroutinesApi
-class CardReaderConnectViewModelTest : BaseUnitTest() {
+class CardReaderDetailViewModelTest : BaseUnitTest() {
     @get:Rule
     val coroutinesTestRule = CoroutineTestRule()
 
-    private lateinit var viewModel: CardReaderConnectViewModel
+    private lateinit var viewModel: CardReaderDetailViewModel
 
     @Before
     fun setUp() {
-        viewModel = CardReaderConnectViewModel(
+        viewModel = CardReaderDetailViewModel(
             SavedStateHandle(),
             coroutinesTestRule.testDispatchers
         )
